lookup from dictionary now supported in text
