+++ index.html
@@ -23,5 +23,6 @@
       </div>
     </form>
     <script src="index.js"></script>
+    <a href="out/test.html">Test</a>
   </body>
 </html>
